backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
pydantic>=2.0.0
python-multipart==0.0.12
feedparser==6.0.11
fuzzywuzzy==0.18.0
